добавил кнопку назад

+++ bot_commans.py
@@ -3,11 +3,12 @@ from functions import *
 from aiogram import types
 from aiogram.dispatcher import FSMContext
 from keyboards.next_chioce import next_choice_buttons
-from keyboards.generated_keyboard import create_keyboard
+from keyboards.generated_keyboard import create_keyboard_with_cancel_button, create_keyboard_with_back_button
 from keyboards.owner import owner_buttons
 from keyboards.back import back_button
 from models import Shop
 from excel_functions import new_doc
+import messages
 
 
 @dp.message_handler(commands="set_commands", state="*")
@@ -88,7 +89,7 @@ async def command_statistics(message: types.Message, state: FSMContext):
             await message.answer('Что то пошло не так, попробуйте позже!', reply_markup=types.ReplyKeyboardRemove())
 
 
-@dp.message_handler(text='🔙 Отмена', state="*")
+@dp.message_handler(text=messages.buttons['cancel'], state="*")
 async def text_cancel_action(message: types.Message, state: FSMContext):
     telegram_id = message.from_user.id
     try:
@@ -117,7 +118,7 @@ async def command_cancel_request(message: types.Message, state: FSMContext):
             employee_requests = get_employee_requests(employee)
             requests = get_accepted_requests(employee_requests)
             shop_names_in_requests = get_datetime_and_shop_names_by_request(requests)
-            requests_keyboard = create_keyboard(shop_names_in_requests)
+            requests_keyboard = create_keyboard_with_cancel_button(shop_names_in_requests)
             await message.answer('Какую заявку отменить?', reply_markup=requests_keyboard)
             await CancelRequest.Cancel.set()
             await state.update_data(shop_names_in_requests=shop_names_in_requests)
@@ -144,7 +145,7 @@ async def command_add_new_shop(message: types.Message):
     employee = get_employee_by_id(telegram_id)
     if employee:
         available_districts = get_the_districts_available_to_the_employee(employee)
-        districts_keyboard = create_keyboard(available_districts)
+        districts_keyboard = create_keyboard_with_cancel_button(available_districts)
         await message.answer('Чтобы добавить новую торговую точку, ответьте на несколько вопросов')
         await message.answer('В каком районе расположена торговая точка?', reply_markup=districts_keyboard)
         await NewShop.District.set()
@@ -254,7 +255,7 @@ async def command_request(message: types.Message, state: FSMContext):
         await state.update_data(employee=employee)
         available_districts = get_the_districts_available_to_the_employee(employee)
         await state.update_data(available_districts=available_districts)
-        districts_keyboard = create_keyboard(available_districts)
+        districts_keyboard = create_keyboard_with_cancel_button(available_districts)
         await message.answer('Выбери район:', reply_markup=districts_keyboard)
 
 
@@ -269,7 +270,7 @@ async def command_request_action_one(message: types.Message, state: FSMContext):
         await state.update_data(all_shops=all_shops)
         shops_names = get_shops_names(all_shops)
         await state.update_data(shops_names=shops_names)
-        shops_keyboard = create_keyboard(shops_names)
+        shops_keyboard = create_keyboard_with_back_button(shops_names)
         await message.answer(f'Выбери название торговой точки:', reply_markup=shops_keyboard)
         await Request.next()
     else:
@@ -280,6 +281,12 @@ async def command_request_action_one(message: types.Message, state: FSMContext):
 async def command_request_action_two(message: types.Message, state: FSMContext):
     shop_name = message.text
     data = await state.get_data()
+    if shop_name == messages.buttons['back']:
+        available_districts = data['available_districts']
+        districts_keyboard = create_keyboard_with_cancel_button(available_districts)
+        await message.answer('Выбери район:', reply_markup=districts_keyboard)
+        await Request.District.set()
+        return
     shops_names = data['shops_names']
     if shop_name in shops_names:
         shop = get_shop(shop_name, data['all_shops'])
@@ -288,7 +295,7 @@ async def command_request_action_two(message: types.Message, state: FSMContext):
         await state.update_data(all_products=all_products)
         products_types = get_products_types(all_products)
         await state.update_data(products_types=products_types)
-        products_types_keyboard = create_keyboard(products_types)
+        products_types_keyboard = create_keyboard_with_back_button(products_types)
         await message.answer(f'Выбери категорию товара: ', reply_markup=products_types_keyboard)
         await Request.next()
     else:
@@ -300,13 +307,19 @@ async def command_request_action_two(message: types.Message, state: FSMContext):
 async def command_request_action_three(message: types.Message, state: FSMContext):
     product_type = message.text
     data = await state.get_data()
+    if product_type == messages.buttons['back']:
+        shops_names = data['shops_names']
+        shops_names_keyboard = create_keyboard_with_back_button(shops_names)
+        await message.answer(f'Выбери название торговой точки:', reply_markup=shops_names_keyboard)
+        await Request.ShopName.set()
+        return
     products_types = data['products_types']
     if product_type in products_types:
         await state.update_data(product_type=product_type)
         products_names = get_products_names_by_type(data['all_products'], product_type)
         await state.update_data(products_names=products_names)
-        products_names_keyboards = create_keyboard(products_names)
-        await message.answer(f'Выбери товар: ', reply_markup=products_names_keyboards)
+        products_names_keyboards = create_keyboard_with_back_button(products_names)
+        await message.answer(f'Выбери товар:', reply_markup=products_names_keyboards)
         await Request.next()
     else:
         await message.answer('Выберите из предложенных, или обратитесь в офис')
@@ -316,6 +329,12 @@ async def command_request_action_three(message: types.Message, state: FSMContext
 async def command_request_action_four(message: types.Message, state: FSMContext):
     product_name = message.text
     data = await state.get_data()
+    if product_name == messages.buttons['back']:
+        products_types = data['products_types']
+        products_types_keyboard = create_keyboard_with_back_button(products_types)
+        await message.answer(f'Выбери категорию товара:', reply_markup=products_types_keyboard)
+        await Request.ProductCategory.set()
+        return
     products_names = data['products_names']
     if product_name in products_names:
         await state.update_data(product_name=product_name)
@@ -331,13 +350,15 @@ async def command_request_action_four(message: types.Message, state: FSMContext)
 
 @dp.message_handler(state=Request.Number, content_types=types.ContentTypes.TEXT)
 async def command_request_action_five(message: types.Message, state: FSMContext):
-    number_of_products = message.text.replace(' ', '')
-    if number_of_products == 'Назад' or int(number_of_products) == 0:
+    number_of_products = message.text
+    if number_of_products == messages.buttons['back'] or int(number_of_products) == 0:
         data = await state.get_data()
-        products_types_keyboard = create_keyboard(data['products_types'])
-        await message.answer(f'Выбери категорию товара: ', reply_markup=products_types_keyboard)
-        await Request.ProductCategory.set()
+        products_names = data['products_names']
+        products_names_keyboard = create_keyboard_with_back_button(products_names)
+        await message.answer(f'Выбери товар:', reply_markup=products_names_keyboard)
+        await Request.Product.set()
         return
+    number_of_products = message.text.replace(' ', '')
     if number_of_products.isdigit():
         data = await state.get_data()
         product = data['product']
@@ -366,7 +387,7 @@ async def command_request_action_five(message: types.Message, state: FSMContext)
         await message.answer(f'Это не число!')
 
 
-@dp.message_handler(text='Добавить товар')
+@dp.message_handler(text=messages.buttons['add'])
 async def text_add_product(message: types.Message, state: FSMContext):
     telegram_id = message.from_user.id
     try:
@@ -380,20 +401,24 @@ async def text_add_product(message: types.Message, state: FSMContext):
     await state.update_data(all_products=all_products)
     products_types = get_products_types(all_products)
     await state.update_data(products_types=products_types)
-    products_types_keyboard = create_keyboard(products_types)
+    products_types_keyboard = create_keyboard_with_back_button(products_types)
     await message.answer(f'Выбери категорию товара: ', reply_markup=products_types_keyboard)
 
 
 @dp.message_handler(state=RequestAdd.ProductCategory, content_types=types.ContentTypes.TEXT)
 async def command_request_add_action_one(message: types.Message, state: FSMContext):
     product_type = message.text
+    if product_type == messages.buttons['back']:
+        await message.answer(f'Что дальше?', reply_markup=next_choice_buttons)
+        await state.finish()
+        return
     data = await state.get_data()
     products_types = data['products_types']
     if product_type in products_types:
         await state.update_data(product_type=product_type)
         products_names = get_products_names_by_type(data['all_products'], product_type)
         await state.update_data(products_names=products_names)
-        products_names_keyboard = create_keyboard(products_names)
+        products_names_keyboard = create_keyboard_with_back_button(products_names)
         await message.answer(f'Выбери товар: ', reply_markup=products_names_keyboard)
         await RequestAdd.next()
     else:
@@ -404,6 +429,12 @@ async def command_request_add_action_one(message: types.Message, state: FSMConte
 async def command_request_add_action_two(message: types.Message, state: FSMContext):
     product_name = message.text
     data = await state.get_data()
+    if product_name == messages.buttons['back']:
+        products_types = data['products_types']
+        products_types_keyboard = create_keyboard_with_back_button(products_types)
+        await message.answer(f'Выбери категорию товара: ', reply_markup=products_types_keyboard)
+        await RequestAdd.ProductCategory.set()
+        return
     products_names = data['products_names']
     if product_name in products_names:
         await state.update_data(product_name=product_name)
@@ -419,13 +450,15 @@ async def command_request_add_action_two(message: types.Message, state: FSMConte
 
 @dp.message_handler(state=RequestAdd.Number, content_types=types.ContentTypes.TEXT)
 async def command_request_add_action_three(message: types.Message, state: FSMContext):
-    number_of_products = message.text.replace(' ', '')
-    if number_of_products == 'Назад' or int(number_of_products) == 0:
+    number_of_products = message.text
+    if number_of_products == messages.buttons['back'] or int(number_of_products) == 0:
         data = await state.get_data()
-        products_types_keyboard = create_keyboard(data['products_types'])
-        await message.answer(f'Выбери категорию товара: ', reply_markup=products_types_keyboard)
-        await RequestAdd.ProductCategory.set()
+        products_names = data['products_names']
+        products_names_keyboard = create_keyboard_with_back_button(products_names)
+        await message.answer(f'Выбери товар: ', reply_markup=products_names_keyboard)
+        await RequestAdd.Product.set()
         return
+    number_of_products = message.text.replace(' ', '')
     if number_of_products.isdigit():
         data = await state.get_data()
         product = data['product']
@@ -461,7 +494,7 @@ async def command_request_add_action_three(message: types.Message, state: FSMCon
         await message.answer(f'Это не число!')
 
 
-@dp.message_handler(text='Удалить товар')
+@dp.message_handler(text=messages.buttons['delete'])
 async def text_delete_product(message: types.Message):
     telegram_id = message.from_user.id
     try:
@@ -470,20 +503,16 @@ async def text_delete_product(message: types.Message):
         await message.answer('Что то пошло не так. Попробуйте заново: /new_request',
                              reply_markup=types.ReplyKeyboardRemove())
         return
-    markup = types.ReplyKeyboardMarkup(resize_keyboard=True, one_time_keyboard=True)
-    for order in orders:
-        button = types.KeyboardButton(order['Номенклатура'])
-        markup.add(button)
-    cancel_button = types.KeyboardButton('Отмена')
-    markup.add(cancel_button)
-    await message.answer(f'Что удалить?', reply_markup=markup)
+    orders_names = [x['Номенклатура'] for x in orders]
+    orders_names_keyboards = create_keyboard_with_back_button(orders_names)
+    await message.answer(f'Что удалить?', reply_markup=orders_names_keyboards)
     await DeleteProduct.Delete.set()
 
 
 @dp.message_handler(state=DeleteProduct.Delete, content_types=types.ContentTypes.TEXT)
 async def text_delete_product_action(message: types.Message, state: FSMContext):
     answer = message.text
-    if answer == 'Отмена':
+    if answer == messages.buttons['back']:
         await message.answer(f'Что дальше?', reply_markup=next_choice_buttons)
         await state.finish()
     else:
@@ -509,7 +538,7 @@ async def text_delete_product_action(message: types.Message, state: FSMContext):
             await message.answer('Товар не найден')
 
 
-@dp.message_handler(text='Завершить 📄')
+@dp.message_handler(text=messages.buttons['end'])
 async def text_the_end(message: types.Message):
     telegram_id = message.from_user.id
     try:
@@ -518,33 +547,36 @@ async def text_the_end(message: types.Message):
         await message.answer('Что то пошло не так. Попробуйте заново: /new_request',
                              reply_markup=types.ReplyKeyboardRemove())
         return
-    request_number = get_last_number_in_requests() + 1
-    datetime_now = time_in_uralsk()
-    request_data = get_product_name_and_count_from_data(data)
-    employee_name = data['employee']['Сокращенное имя']
-    request = [[request_number, datetime_now, data['shop'].get('Название', 'Неизвестно'),
-                data['shop'].get('ИП/ТОО', 'Неизвестно'), data['shop'].get('Адрес', 'Неизвестно'),
-                request_data, data['total_sum'], data['shop'].get('Кассовый аппарат', 'Неизвестно'),
-                employee_name, 'Принят']]
-    append_data_in_table(config.REQUESTS, list_name='Все', user_value=request)
-    lists_in_table = get_lists_names_in_table(config.REQUESTS)
-    if employee_name not in lists_in_table:
-        create_new_list_in_table(config.REQUESTS, employee_name)
-        add_base_titles_from_the_first_page_in_list(config.REQUESTS, employee_name)
-    append_data_in_table(config.REQUESTS, list_name=employee_name, user_value=request)
-    add_link_to_request_status(employee_name)
-    uralsk_date = time_in_uralsk_date()
-    new_doc(data, request_number, uralsk_date)
-    file_name = f'Счет-фактура {request_number}.xlsx'
-    try:
-        with open(r'./docs/' + file_name, 'rb') as document:
-            await message.answer_document(document)
-            delete_file(file_name)
-    except:
-        pass
-    finally:
-        delete_json_file(telegram_id)
-    await message.answer('Заявка была принята!', reply_markup=types.ReplyKeyboardRemove())
+    if data["orders"]:
+        request_number = get_last_number_in_requests() + 1
+        datetime_now = time_in_uralsk()
+        request_data = get_product_name_and_count_from_data(data)
+        employee_name = data['employee']['Сокращенное имя']
+        request = [[request_number, datetime_now, data['shop'].get('Название', 'Неизвестно'),
+                    data['shop'].get('ИП/ТОО', 'Неизвестно'), data['shop'].get('Адрес', 'Неизвестно'),
+                    request_data, data['total_sum'], data['shop'].get('Кассовый аппарат', 'Неизвестно'),
+                    employee_name, 'Принят']]
+        append_data_in_table(config.REQUESTS, list_name='Все', user_value=request)
+        lists_in_table = get_lists_names_in_table(config.REQUESTS)
+        if employee_name not in lists_in_table:
+            create_new_list_in_table(config.REQUESTS, employee_name)
+            add_base_titles_from_the_first_page_in_list(config.REQUESTS, employee_name)
+        append_data_in_table(config.REQUESTS, list_name=employee_name, user_value=request)
+        add_link_to_request_status(employee_name)
+        uralsk_date = time_in_uralsk_date()
+        new_doc(data, request_number, uralsk_date)
+        file_name = f'Счет-фактура {request_number}.xlsx'
+        try:
+            with open(r'./docs/' + file_name, 'rb') as document:
+                await message.answer_document(document)
+                delete_file(file_name)
+        except:
+            pass
+        finally:
+            delete_json_file(telegram_id)
+        await message.answer('Заявка была принята!', reply_markup=types.ReplyKeyboardRemove())
+    else:
+        await message.answer('Не могу добавить заявку. Она пустая!')
 
 
 @dp.message_handler()
